Set up per-compilation directories.
This commit gives the opportunity for the user
to take advantage of the context (i.e. which
compilation is the current one) in the options
which are in custom.mk (or in flit-config.toml).

There are two parts of this:
  * Declare a CUR_DIR variable and make it
    available anywhere there is a compilation rule.
    This can then be used by, for example, in custom.mk
    or in the toml file. Also, use a hack to create
    the appropriate directory if it does not exist.
  * Change variables like CC_REQUIRED and F_REQUIRED to
    be lazy variables. Without this, the rules would be
    evaluated too early for those places to be used by
    custom.mk or toml options.

I understand that eager variables are generally preferred,
but I think this is the exact use case for lazy variables.

Overally, there is a substantial bit of hackery to make everything
work. In the future, a more general way to do this would be to
create a canonical obj/ and include/ directory per test. However,
right now I don't want to break and fix the object files.