Use directories instead of suffixes
Previously, FLiT has used suffixes in object files
to distinguish between different compilations.
This commit changes that to use directory prefixes
instead. Now, object files for the compilation XXX
can be found in the directory obj/XXX/.

The previous $(CUR_DIR) feature (which happened to
be a naming conflict) from the last commit has
been modified so that the suffix "_include" is added
to this directory. This is a bit hacky, though, and
should be re-thought in the future.